NN-2177 fix integration tests
