+++ src/test/kotlin/uk/gov/justice/digital/hmpps/whereabouts/integration/IntegrationTest.kt
@@ -2,6 +2,8 @@ package uk.gov.justice.digital.hmpps.whereabouts.integration
 
 import com.google.gson.Gson
 import groovy.util.logging.Slf4j
+import org.junit.Before
+import org.junit.ClassRule
 import org.junit.runner.RunWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
@@ -18,10 +20,14 @@ import org.springframework.test.context.ContextConfiguration
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner
 import uk.gov.justice.digital.hmpps.whereabouts.common.getGson
 import uk.gov.justice.digital.hmpps.whereabouts.dto.AttendanceDto
+import uk.gov.justice.digital.hmpps.whereabouts.integration.wiremock.CaseNotesMockServer
+import uk.gov.justice.digital.hmpps.whereabouts.integration.wiremock.Elite2MockServer
+import uk.gov.justice.digital.hmpps.whereabouts.integration.wiremock.OAuthMockServer
 
 class ListOfAttendanceDtoReferenceType : ParameterizedTypeReference<List<AttendanceDto>>()
 
 @RunWith(SpringJUnit4ClassRunner::class)
+//@RunWith(SpringRunner::class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
 @ContextConfiguration
@@ -35,8 +41,33 @@ abstract class IntegrationTest {
     @Value("\${token}")
     private val token: String? = null
 
+    companion object {
+        @get:ClassRule
+        @JvmStatic
+        val elite2MockServer = Elite2MockServer()
+
+        @get:ClassRule
+        @JvmStatic
+        val oauthMockServer = OAuthMockServer()
+
+        @get:ClassRule
+        @JvmStatic
+        val caseNotesMockServer = CaseNotesMockServer()
+    }
+
     init {
         SecurityContextHolder.getContext().authentication = TestingAuthenticationToken("user", "pw")
+        // Resolves an issue where Wiremock keeps previous sockets open from other tests causing connection resets
+        System.setProperty("http.keepAlive", "false");
+    }
+
+    @Before
+    fun resetStubs() {
+        elite2MockServer.resetAll()
+        oauthMockServer.resetAll()
+        caseNotesMockServer.resetAll()
+
+        oauthMockServer.stubGrantToken()
     }
 
     internal fun createHeaderEntity(entity: Any): HttpEntity<*> {
